+++ lib/services/test-execution-service.ts
@@ -152,7 +152,7 @@ export class TestExecutionService implements ITestExecutionService {
 			paths: [projectData.projectDir],
 		});
 
-		canStartKarmaServer = !!pathToKarma;
+		canStartKarmaServer = canStartKarmaServer && !!pathToKarma;
 
 		return canStartKarmaServer;
 	}
@@ -190,6 +190,7 @@ export class TestExecutionService implements ITestExecutionService {
 				path: this.$options.path,
 				tns: process.argv[1],
 				node: process.execPath,
+				env: this.$options.env,
 				options: {
 					debugTransport: this.$options.debugTransport,
 					debugBrk: this.$options.debugBrk,